kids, remember to sanitize your inputs

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: ylabussi <ylabussi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 21:07:06 by ylabussi          #+#    #+#             */
-/*   Updated: 2024/12/12 21:17:54 by ylabussi         ###   ########.fr       */
+/*   Updated: 2025/05/14 14:43:12 by ylabussi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_calloc(size_t nmemb, size_t size)
 {
 	void	*p;
 
-	if (nmemb == 0 || size > ((size_t) -1 / nmemb))
+	if (nmemb == 0 || size > (-1ul / nmemb))
 		return (NULL);
 	p = malloc(nmemb * size);
 	if (!p)
